\name{utInit}
\alias{utInit}
\title{Initialize Udunits Library}
\description{
 Initializes the udunits library.  This function must be called before calling
 any other of the udunits functions.
}
\usage{
 utInit()
}
\references{
\url{http://www.unidata.ucar.edu/packages/udunits/}
}
\author{Library routines by Unidata; interface glue by David W. Pierce \email{dpierce@ucsd.edu}}
\seealso{ \code{\link[udunits]{utScan}}, 
 \code{\link[udunits]{utCalendar}}, \code{\link[udunits]{utInvCalendar}},
 \code{\link[udunits]{utFormatDate}}, \code{\link[udunits]{utDayOfWeek}}, \code{\link[udunits]{utIsTime}},
 \code{\link[udunits]{utHasOrigin}}, \code{\link[udunits]{utConvert}} }
\keyword{utilities}
