\name{version.udunits}
\alias{version.udunits}
\title{Prints the version of the udunits library}
\description{
 Prints the version of the udunits library.
}
\usage{
 version.udunits()
}
\arguments{
}
\value{A character string, such as "1.2", that indicates the version of the udunits library.}
\references{
\url{http://www.unidata.ucar.edu/packages/udunits/}
}
\author{Library routines by Unidata; interface glue by David W. Pierce \email{dpierce@ucsd.edu}}
\keyword{utilities}
